bitkeeper revision 1.1336 (42662e0bB62AMXvz0XCbYomA5_OHAw)
[PATCH] x86-64-enable-vmx.patch

   Enable CONFIG_VMX for x86_64.

   - Provides vmexit/entry handling code based on the x86_32 code
   - Fix find_highest_vector for 64 bit (Benjamin Liu)

+++ xen/arch/x86/vmx_intercept.c
@@ -37,7 +37,7 @@ int vmx_io_intercept(ioreq_t *p)
     struct exec_domain *d = current;
     struct vmx_handler_t *handler = &(d->arch.arch_vmx.vmx_platform.vmx_handler);
     int i;
-    unsigned addr, offset;
+    unsigned long addr, offset;
     for (i = 0; i < handler->num_slot; i++) {
         addr   = handler->hdl_list[i].addr;
         offset = handler->hdl_list[i].offset;
